requirements.txt
aniso8601==6.0.0
appdirs==1.4.3
APScheduler==3.5.3
attrs==19.3.0
baseball==1.4
certifi==2019.3.9
chardet==3.0.4
Click==7.0
distlib==0.3.0
filelock==3.0.12
Flask==1.0.2
flask-restplus==0.13.0
idna==2.8
importlib-metadata==1.5.0
importlib-resources==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==3.2.0
MarkupSafe==1.1.1
ptvsd==4.3.2
PyJWT==1.7.1
PyMySQL==0.9.3
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2018.9
requests==2.21.0
six==1.12.0
tzlocal==1.5.1
urllib3==1.24.1
uWSGI==2.0.18
virtualenv==20.0.2
Werkzeug==0.14.1
zipp==2.1.0
